+++ Libs/Indicators/Outputs/PointOfTime.cs
@@ -69,5 +69,5 @@ public class PointOfTime
     /// 当期收盘价
     /// </summary>
     [JsonProperty("close", NullValueHandling = NullValueHandling.Ignore)]
-    public decimal Close { get; set; }
+    public double Close { get; set; }
 }
