+++ Libs/Indicators/Outputs/StockLog.cs
@@ -40,5 +40,5 @@ public class StockLog
     /// 收盘价
     /// </summary>
     [JsonProperty("close", NullValueHandling = NullValueHandling.Ignore)]
-    public decimal Close { get; set; }
+    public double Close { get; set; }
 }
